* http://otfans.net/project.php?issueid=688 * http://otfans.net/project.php?issueid=685

+++ otserv/trunk/movement.cpp
@@ -775,7 +775,7 @@ uint32_t MoveEvent::EquipItem(MoveEvent* moveEvent, Player* player, Item* item, 
 	//This includes item transforming
 	if(!player->hasFlag(PlayerFlag_IgnoreWeaponCheck) && moveEvent->getWieldInfo() != 0){
 		if((int32_t)player->getLevel() < moveEvent->getReqLevel()
-			|| (int32_t)player->getMagicLevel() < moveEvent->getReqMagLv() ||
+			|| player->getMagicLevel() < moveEvent->getReqMagLv() ||
 			(!player->isPremium() && moveEvent->isPremium())){
 			return 1;
 		}
@@ -1094,7 +1094,7 @@ ReturnValue MoveEvent::canPlayerWearEquip(Player* player, slots_t slot)
 	if((int32_t)player->getLevel() < getReqLevel()){
 		return RET_NOTREQUIREDLEVEL;
 	}
-	if((int32_t)player->getMagicLevel() < getReqMagLv()){
+	if(player->getMagicLevel() < getReqMagLv()){
 		return RET_NOTENOUGHMAGICLEVEL;
 	}
 	if(!player->isPremium() && isPremium()){
