+++ otserv/trunk/spells.cpp
@@ -646,7 +646,7 @@ bool Spell::playerSpellCheck(Player* player) const
 			return false;
 		}
 
-		if(((int32_t)player->getMagicLevel()) < magLevel){
+		if((player->getMagicLevel()) < magLevel){
 			player->sendCancelMessage(RET_NOTENOUGHMAGICLEVEL);
 			g_game.addMagicEffect(player->getPosition(), NM_ME_PUFF);
 			return false;
